Blacklist based on request fingerprint.
